f5mcp/requirements.txt
fastmcp
bigrest
requests
python-dotenv
